requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions>=1.12.0
fastapi==0.97.0
pandas==2.0.2
pydantic==1.10.9
requests==2.31.0
python-dotenv==1.0.0
azure-core==1.29.3